% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{AT}
\alias{AT}
\title{Adams' Thesis data}
\format{
A data frame with 4,590 rows and 5 variables:
\describe{
\item{submission_id}{A unique identifier for each participant.}
\item{item_nr}{An identifier for the context-statement pair presented
to the participant.}
\item{rating}{The kind of rating the participant has given.}
\item{cond_type}{The type of conditional statement.}
\item{response}{The measured acceptability rating on a 7-point Likert scale.}
}
}
\usage{
AT
}
\description{
Data from a replication of Douven and Verbrugge (2010)
investigating the relationship between acceptability ratings of conditional
sentences and conditional probabilities (variable \code{rating}) across
three different types of conditionals (inductive, abductive, deductive).
}
\examples{
\dontrun{
# fit an ordinal model using 'brms'
# regress response as a function of rating and conditional type (with interaction)
fit <- brms::brm(response ~ rating * cond_type,
                 data = AT,
                 family = cumulative("logit"))

# check encoding scheme used in the model
get_cell_definitions(fit)

# get draws for acceptability averaged over the type of conditional
extract_cell_draws(fit, rating == "acceptability")

# compare main effects of rating
compare_groups(fit, rating == "acceptability", rating == "cond_probability")
}

}
\references{
Douven, I., Verbrugge, S. (2010). The Adams Family. \emph{Cognition},
\emph{117}, 302-18. \url{https://doi.org/10.1016/j.cognition.2010.08.015}.
}
\keyword{datasets}
