% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare means of two subsets of factorial design cells}
\usage{
compare_groups(model, higher, lower)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
list with posterior samples for each group, and the posterior probability that group 'higher' has a higher estimated coefficient in the posterior samples than the group 'lower'
}
\description{
This function takes a brms model fit for a factorial design and a specification of two groups (subsets of design cells) to compare. 
A group is specified as a named list, specifiying the factors and their levels which to include in the group.
It outputs the posterior mean of the 'higher' minus the 'lower' subset of cells, its 95 percent credible interval and the posterior probability that the 'higher' group has a higher mean than the the 'lower' group.
For more information see \code{vignette('faintr_basics')}.
}
\examples{
library(brms)
m = brm(yield ~ N * P * K, npk)
# this compares two single cells in the factorial design
compare_groups(
 model = m, 
 higher = list("N" = "1", "P" = "1",  "K" = "1"), 
 lower  = list("N" = "0", "P" = "0",  "K" = "1")
)
# this compares the average of N=1 cells to the grand mean
# like in deviance conding
compare_groups(
 model = m, 
 higher = list("N" = "1"), 
 lower  = list()
)

}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
