% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{extract_cell_draws}
\alias{extract_cell_draws}
\title{Extract posterior draws for one subset of factorial design cells}
\usage{
extract_cell_draws(fit, group, colname = "draws")
}
\arguments{
\item{fit}{An object of class \code{\link[brms]{brmsfit}}.}

\item{group}{An expression specifying the group to filter the draws for.}

\item{colname}{A string specifying the column name of the returned data frame;
defaults to 'draws'.}
}
\value{
A data frame containing posterior draws for the specified group.
}
\description{
This function takes as input a \code{\link[brms]{brms}} model fit for a
factorial design and a group specification (one subset of the design cells)
and returns the posterior draws for that group. If no group is specified,
the returned draws are grand means.
}
\note{
The \pkg{faintr} package currently does not support multivariate models and
models that use families \code{categorical}, \code{dirichlet}, \code{multinomial},
and \code{logistic_normal}. Furthermore, models must not include special effect
terms \code{mo()}, \code{mi()}, \code{me()}, and \code{cs()} for fixed effects.
Also note that \pkg{faintr} currently does not support models where the intercept
is a population-level parameter (class \code{b}), as is the case when using the
\code{0 + Intercept} syntax in the \code{brm} function call.
}
\examples{
\dontrun{
# fit a linear mixed effects model
# regressing voice pitch against gender and context and including random effects
fit <- brms::brm(formula = pitch ~ gender * context + (1 | subject + sentence),
                 data = politeness)

# extract draws for female speakers in informal contexts
extract_cell_draws(fit, gender == "F" & context == "inf")

# extract draws for male speakers or informal contexts
extract_cell_draws(fit, gender == "M" | context == "inf")

# averaged over gender, extract draws for all but polite contexts
extract_cell_draws(fit, context != "pol")

# extract posterior draws averaged over all factors (grand mean)
extract_cell_draws(fit)
}

}
\references{
Bürkner, P.-C. (2017). brms: An R Package for Bayesian Multilevel Models
Using Stan. \emph{Journal of Statistical Software}, \emph{80}(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
