% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{extract_posterior_cell_means}
\alias{extract_posterior_cell_means}
\title{Extracting cell means}
\usage{
extract_posterior_cell_means(model)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\description{
This function takes a brms model fit for a factorial design and outputs a comparison of all factor levels.
}
\examples{
extract_posterior_cell_means()
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
