\name{faintr-package}
\alias{faintr-package}
\alias{faintr}
\docType{package}
\title{
\packageTitle{faintr}
}
\description{
\packageDescription{The \pkg{faintr} package provides convenience functions for testing different hypotheses about factor level combinations for a Bayesian regression analysis of data from a factorial-design experiment. It builds on a regression model fitted with the help of the \pkg{brms} package. For more information type \code{vignette("faintr_basics")}. }
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{faintr}
\packageIndices{faintr}
Functions provided by this packages are: \code{\link{compare_groups}}. 
}
\author{
\packageAuthor{Michael Franke}
Maintainer: \packageMaintainer{Michael Franke}
}
\keyword{ 
factorial design
dummy coding
Bayesian regression modeling
}
\seealso{
\code{\link[brms:brms-package]{brms}}
}
\examples{
See vignette by typing \code{vignette("faintr_basics")}.
}
