% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{get_cell_comparison}
\alias{get_cell_comparison}
\title{Comparing cell means}
\usage{
get_cell_comparison(model, higher, lower)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
list with posterior samples for each group, and the posterior probability that group 'higher' has a higher estimated coefficient in the posterior samples than the group 'lower'
}
\description{
This function takes a brms model fit and design cells to compare and it outputs the posterior probability that the 'higher' cell has a higher coefficient than the 'lower' group.
}
\examples{
library(brms)
m = brm(yield ~ N * P * K, npk)
get_cell_comparison(
 model = m, 
 higher = list("N" = "1", "P" = "1",  "K" = "1"), 
 lower  = list("N" = "0", "P" = "0",  "K" = "1")
)
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
