% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{get_cell_definitions}
\alias{get_cell_definitions}
\title{Obtain information about factors in regression models}
\usage{
get_cell_definitions(fit)
}
\arguments{
\item{fit}{An object of class \code{\link[brms]{brmsfit}}.}
}
\value{
A \code{\link[tibble]{tibble}} containing the minimal design matrix.
}
\description{
This function takes a \code{\link[brms]{brms}} model fit for a
factorial design and returns all design cells along with their encoding in
the regression model.
}
\note{
The \pkg{faintr} package currently does not support multivariate models and
models that use families \code{categorical}, \code{dirichlet}, \code{multinomial},
and \code{logistic_normal}. Furthermore, models must not include special effect
terms \code{mo()}, \code{mi()}, \code{me()}, and \code{cs()} for fixed effects.
Also note that \pkg{faintr} currently does not support models where the intercept
is a population-level parameter (class \code{b}), as is the case when using the
\code{0 + Intercept} syntax in the \code{brm} function call.
}
\examples{
\dontrun{
# fit a linear mixed effects model on data from a 2 x 2 factorial design
## regress voice pitch against gender and context
fit <- brms::brm(formula = pitch ~ gender * context + (1 | subject + sentence),
                 data = politeness)

# check encoding scheme used in the model
get_cell_definitions(fit)
}

}
\references{
Bürkner, P.-C. (2017). brms: An R Package for Bayesian Multilevel Models
Using Stan. \emph{Journal of Statistical Software}, \emph{80}(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
