% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{get_factor_information}
\alias{get_factor_information}
\title{Obtaining information about factors in regression model}
\usage{
get_factor_information(model)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
list with names of factors and their levels, including the reference levels (in dummy coding)
}
\description{
For a model for a factorial design, fitted with brms, this function returns information about the factors used, their levels, and the reference levels.
For more information see \code{vignette('faintr_basics')}.
}
\examples{
library(brms)
m = brm(yield ~ N * P * K, npk)
get_factor_information(m)
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
