% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{MC}
\alias{MC}
\title{Mental Chronometry data}
\format{
A data frame with 2,404 rows and 4 variables:
\describe{
\item{submission_id}{A unique identifier for each participant.}
\item{shape}{The shape presented on the screen.}
\item{block}{The task of how to react to a shape.}
\item{RT}{The measured reaction time in milliseconds.}
}
}
\usage{
MC
}
\description{
Data from an experiment investigating differences in reaction times
across three tasks of increasing complexity.

Participants reacted to a square or a circle (variable \code{shape})
appearing on the screen by
\itemize{
\item pressing the space bar whenever they saw any shape (block \code{reaction}).
\item pressing the space bar whenever they saw a specific shape (block \code{goNoGo}).
\item pressing a key when they saw one of the shapes and a different key when they
saw the other one (block \code{discrimination}).
}
}
\examples{
\dontrun{
# fit a linear model using 'brms'
# regress reaction time as a function of the block
fit <- brms::brm(RT ~ block, data = MC)

# check encoding scheme used in the model
get_cell_definitions(fit)

# get draws for the reaction block
extract_cell_draws(fit, block == "reaction")

# compare the goNoGo block with the reaction block
compare_groups(fit, block == "goNoGo", block == "reaction")
}

}
\keyword{datasets}
