% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{print.faintCompare}
\alias{print.faintCompare}
\title{Print comparison object between factor groups}
\usage{
\method{print}{faintCompare}(obj)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
string
}
\description{
Print comparison object between factor groups
}
\examples{
print(model_fit)
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
