% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare two subsets of factorial design cells}
\usage{
compare_groups(fit, higher, lower, hdi = 0.95)
}
\arguments{
\item{fit}{An object of class \code{\link[brms]{brmsfit}}.}

\item{higher}{An expression specifying the 'higher' group to filter the draws for.}

\item{lower}{An expression specifying the 'lower' group to filter the draws for.}

\item{hdi}{A single value (0, 1) defining the probability mass within the
highest density interval; defaults to 0.95.}
}
\value{
An object of class 'faintCompare' containing summary statistics of the comparison.
}
\description{
Convenience function for comparing two subsets of factorial design cells.
It takes a \code{\link[brms]{brms}} model fit, two group specifications
(two subsets of design cells), and the probability mass within the highest
density interval of the difference in groups. It outputs the posterior mean
of the 'higher' minus the 'lower' group, a credible interval of the mean difference,
as well as the posterior probability and odds that the mean estimate of the 'higher'
group is higher than that of the 'lower' group. A comparison of one group against
the grand mean can be obtained by leaving out one of the two group specifications
in the function call.
}
\note{
The \pkg{faintr} package currently does not support multivariate models and
models that use families \code{categorical}, \code{dirichlet}, \code{multinomial},
and \code{logistic_normal}. Furthermore, models must not include special effect
terms \code{mo()}, \code{mi()}, \code{me()}, and \code{cs()} for fixed effects.
Also note that \pkg{faintr} currently does not support models where the intercept
is a population-level parameter (class \code{b}), as is the case when using the
\code{0 + Intercept} syntax in the \code{brm} function call.
}
\examples{
\dontrun{
# fit a linear mixed effects model
# regressing voice pitch against gender and context and including random effects
fit <- brms::brm(formula = pitch ~ gender * context + (1 | subject + sentence),
                 data = politeness)

# compare female speakers in informal contexts against male speakers in polite contexts
compare_groups(
 fit  = fit,
 higher = gender == "F" & context == "inf",
 lower  = gender == "M" & context == "pol"
)

# compare informal contexts against polite contexts, averaged over gender
compare_groups(
 fit  = fit,
 higher = context == "inf",
 lower  = context == "pol"
)

# compare female speakers against the grand mean
compare_groups(
 fit  = fit,
 higher = gender == "F",
 hdi = 0.8
)
}

}
\references{
Bürkner, P.-C. (2017). brms: An R Package for Bayesian Multilevel Models
Using Stan. \emph{Journal of Statistical Software}, \emph{80}(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
