% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{extract_posterior_cell_means}
\alias{extract_posterior_cell_means}
\title{Extracting cell means}
\usage{
extract_posterior_cell_means(model)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
list with (i) samples of estimated means of all cells, (ii) pairwise comparison of each cell (whether one has credibly a higher inferred mean than the other), and (iii) a summary (mean & 95% HDI) for each posterior estimate of cell means.
}
\description{
This function takes a brms model fit for a factorial design and outputs a comparison of all factor levels against each other, and posterior samples of all cell means. 
For more information see \code{vignette('faintr_basics')}.
}
\examples{
#' library(brms)
m = brm(yield ~ N * P * K, npk)
extract_posterior_cell_means(m)
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
