% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr-package.R
\docType{package}
\name{faintr-package}
\alias{faintr-package}
\alias{faintr}
\title{Factorial-design interpreter for Bayesian regression models}
\description{
he \pkg{faintr} package provides convenience functions for testing different hypotheses 
about factor level combinations for a Bayesian regression analysis of data from a 
factorial-design experiment. It builds on a regression model fitted with the help of 
the \pkg{brms} package. For more information see \href{https://michael-franke.github.io/bayes_mixed_regression_tutorial/faintr/doc/faintr_basics.html}{the basic tutorial online}.
}
\details{
The package provides the following functions.
\code{\link{get_factor_information}} extracts information about the factors
and their (reference) levels used in the regression analysis.
\code{\link{post_cells}} gives the estimated means for all 
design cells, and a complete pairwise comparison between cells.
\code{\link{compare_groups}} allows the comparison of any pair of 
(subsets) of cells.

For more information see \href{https://michael-franke.github.io/bayes_mixed_regression_tutorial/faintr/doc/faintr_basics.html}{the basic tutorial online}.
}
\seealso{
\code{\link{brm}}, 
\code{\link{brmsfit}}
}
