% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{get_cell_comparison}
\alias{get_cell_comparison}
\title{Comparing cell means}
\usage{
get_cell_comparison(model, cell_low, cell_high)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\description{
This function takes a brms model fit and design cells to compare and it outputs posterior probabilities over differences between cell means.
}
\examples{
get_cell_comparison()
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
