% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{print.faintCompare}
\alias{print.faintCompare}
\title{Print group comparison object}
\usage{
\method{print}{faintCompare}(x, ...)
}
\arguments{
\item{x}{An object containing the summary statistics of a group comparison.}

\item{...}{Ignored.}
}
\description{
Print group comparison object
}
